Move files to buildkite/collector namespace
