+++ lib/buildkite/collector/library_hooks/rspec.rb
