+++ lib/buildkite/collector/object.rb
