+++ lib/buildkite/collector/rspec_plugin/reporter.rb
