+++ lib/buildkite/collector/rspec_plugin/trace.rb
