+++ lib/buildkite/collector/session.rb
