+++ lib/buildkite/collector/socket_connection.rb
