+++ lib/buildkite/collector/tracer.rb
