+++ lib/buildkite/collector/uploader.rb
